[code quality] ~ fix some code quality issues [code quality] ~ fix some code quality issues

+++ services/poker_game/include/Round.hpp
@@ -8,7 +8,7 @@
 class Player;
 
 class Round {
-        enum class Street : int {
+        enum class Street : int8_t {
             PREFLOP = 0,
             FLOP,
             TURN,
